+++ tutorials/plot_object_epochs.py
@@ -137,4 +137,7 @@ _ = ev_right.plot(axes=axs[:, 1], show=False)
 plt.tight_layout()
 
 ###############################################################################
-# To export and manipulate Epochs using Pandas see :ref:`tut_io_export_pandas`.
+# To export and manipulate Epochs using Pandas see
+# :ref:`sphx_glr_auto_tutorials_plot_epochs_to_data_frame.py`,
+# or to work directly with metadata in MNE-Python see
+# :ref:`sphx_glr_auto_tutorials_plot_metadata_epochs.py`.