+++ doc/whats_new.rst
@@ -19,7 +19,7 @@ Current
 Changelog
 ~~~~~~~~~
 
-- Nothing yet
+- Add support for metadata in :class:`mne.Epochs` by `Chris Holdgraf`_, `Alex Gramfort`_, `Jona Sassenhagen`_, and `Eric Larson`_
 
 Bug
 ~~~